+++ front/src/views/Home.vue
@@ -18,4 +18,4 @@ export default {
             appStore.getDadosCep[item] }}</li>
     </ul>
 </template>
-<style></style>
+<style></style>@/pinia/appStore
